requirements/base.txt
amqp==2.5.0
ansible==2.8.2
anyjson==0.3.3
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
billiard==3.6.0.0
boto3==1.9.188
botocore==1.12.188
celery==4.3.0
cffi==1.12.3
cryptography==2.7
Django==2.2.3
django-cors-headers==2.4.0
djangorestframework==3.9.4
docutils==0.14
gunicorn==19.9.0
importlib-metadata==0.18
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.6.3
MarkupSafe==1.1.1
more-itertools==7.1.0
packaging==19.0
pluggy==0.12.0
psycopg2==2.8.3
py==1.8.0
pycparser==2.19
pyparsing==2.4.0
pytest==5.0.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
redis==3.2.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
zipp==0.5.2
